+++ lib/model/entry/entry.dart
@@ -491,7 +491,8 @@ class AvesEntry with AvesEntryBase {
     bool oldIsFlipped,
   ) async {
     if ((!MimeTypes.refersToSameType(oldMimeType, mimeType) && !MimeTypes.isVideo(oldMimeType)) || oldDateModifiedMillis != dateModifiedMillis || oldRotationDegrees != rotationDegrees || oldIsFlipped != isFlipped) {
-      await EntryCache.evict(uri, oldMimeType, oldDateModifiedMillis, oldRotationDegrees, oldIsFlipped, isAnimated);
+      EntryCache.evict(uri);
+      await mediaFetchService.clearDecoders();
       visualChangeNotifier.notify();
     }
   }