+++ lib/widgets/navigation/drawer/app_drawer.dart
@@ -104,15 +104,10 @@ class _AppDrawerState extends State<AppDrawer> with WidgetsBindingObserver {
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
-    switch (state) {
-      case .resumed:
-        if (_profileSwitchPermissionRequested) {
-          _profileSwitchPermissionRequested = false;
-          _initProfileSwitchFuture();
-          setState(() {});
-        }
-      default:
-        break;
+    if (state == .resumed && _profileSwitchPermissionRequested) {
+      _profileSwitchPermissionRequested = false;
+      _initProfileSwitchFuture();
+      setState(() {});
     }
   }