+++ lib/widgets/viewer/visual/video/cover.dart
@@ -59,7 +59,7 @@ class _VideoCoverState extends State<VideoCover> {
   Size get videoDisplaySize => widget.videoDisplaySize;
 
   // use the high res photo as cover for the video part of a motion photo
-  ImageProvider get videoCoverUriImage => (mainEntry.isMotionPhoto ? mainEntry : entry).fullImage;
+  ImageProvider get videoCoverUriImage => (mainEntry.isMotionPhoto ? mainEntry : entry).getFullImage();
 
   @override
   void initState() {